+++ NelmioSecurityBundle.php
@@ -12,7 +12,15 @@
 namespace Nelmio\SecurityBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Nelmio\SecurityBundle\DependencyInjection\Compiler\CSPTwigCompilerPass;
 
 class NelmioSecurityBundle extends Bundle
 {
+    public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $container->addCompilerPass(new CSPTwigCompilerPass());
+    }
 }